alpha_factory_v1/requirements-lock.txt
#
#  ──────────────────────────────────────────────────────────────
#   Alpha‑Factory • dependency lock – generated 2025‑04‑20
#   Platform:  Python 3.11  •  OS/Arch: manylinux_x86_64
#  ──────────────────────────────────────────────────────────────
#
accelerate==0.27.2             # via -r backend/requirements.txt
aiofiles==23.2.1               # transitive (starlette upload)
aiohttp==3.9.4                 # uvicorn[standard] websockets
aiosignal==1.3.1               # aiohttp
anyio==4.3.0                   # httpx, starlette
anthropic==0.21.0              # via -r backend/requirements.txt
async-timeout==4.0.3           # aiohttp
better-profanity==0.7.0        # via -r backend/requirements.txt
certifi==2024.2.2              # requests / httpx
charset-normalizer==3.4.1      # requests
chromadb==0.5.23               # via -r backend/requirements.txt
click==8.1.7                   # gunicorn, uvicorn
dataclasses-json==0.6.5        # sentence-transformers
diskcache==5.6.3               # chromadb
faiss-cpu==1.7.4               # via -r backend/requirements.txt
fastapi==0.110.0               # via -r backend/requirements.txt
filetype==1.2.0                # sentence-transformers
frozenlist==1.4.1              # aiohttp
fsspec==2024.3.1               # sentence-transformers
gunicorn==22.0.0               # via -r backend/requirements.txt
h11==0.14.0                    # httpcore
httpcore==1.0.4                # httpx
httptools==0.6.1               # uvicorn[standard]
httpx==0.27.0                  # openai, anthropic
idna==3.7                      # requests, httpx
jinja2==3.1.3                  # rocketry (templates)
joblib==1.4.2                  # sentence-transformers
litellm==1.31.0                # via -r backend/requirements.txt
markupsafe==2.1.5              # jinja2
nltk==3.8.1                    # sentence-transformers
numpy==1.26.4                  # via -r backend/requirements.txt
openai==1.14.0                 # via -r backend/requirements.txt
openai-agents==0.0.13          # via -r backend/requirements.txt
orjson==3.10.1                 # via -r backend/requirements.txt
ortools==9.7.2996              # via -r backend/requirements.txt
packaging==23.2                # accelerate, transformers
pandas==2.2.1                  # via -r backend/requirements.txt
pillow==10.3.0                 # sentence-transformers
prometheus-client==0.19.0      # via -r backend/requirements.txt
psutil==5.9.8                  # accelerate, rocketry
pydantic==2.5.3                # via -r backend/requirements.txt, fastapi
pydantic-core==2.16.3          # pydantic
python-dateutil==2.9.0.post0   # pandas
python-dotenv==1.0.1           # rocketry convenience
pytz==2024.1                   # pandas
pyyaml==6.0.1                  # accelerate
regex==2025.3.25               # sentence-transformers, transformers
rocketry==2.5.1                # via -r backend/requirements.txt
scikit-learn==1.5.0            # sentence-transformers
scipy==1.12.0                  # via -r backend/requirements.txt
sentence-transformers==2.4.0   # via -r backend/requirements.txt
sentencepiece==0.1.99          # via -r backend/requirements.txt
setproctitle==1.3.3            # gunicorn
six==1.16.0                    # python-dateutil, nltk
sniffio==1.3.1                 # anyio
starlette==0.36.3              # fastapi
threadpoolctl==3.5.0           # scikit-learn
tiktoken==0.5.0                # via -r backend/requirements.txt
tokenizers==0.19.1             # transformers, chromadb
torch==2.3.0                   # accelerate, sentence-transformers
transformers==0.20.0           # via -r backend/requirements.txt
typing-extensions==4.11.0      # pydantic, transformers
urllib3==2.2.2                 # requests, httpx
uvicorn[standard]==0.29.0      # via -r backend/requirements.txt
uvloop==0.19.0                 # uvicorn[standard]
watchfiles==0.21.0             # uvicorn[standard]
websockets==12.0               # uvicorn[standard]

# --------------------------- END OF FILE ---------------------------
